slides: migrate coverage_bitmap* / text rasterization to u8 storage
Text coverage buffers already only held 0-255.  Switch the storage
type from uint16_t to uint8_t and the samplers in slides/coverage.c
from umbra_load_16 / umbra_gather_16 + i32_from_s16 over to the new
umbra_load_8 / umbra_gather_8.  Halves the text atlas footprint and
drops the widening op from every coverage lookup.

Propagates to tests/draw_test.c and tests/golden_test.c whose
coverage_bitmap* samplers bind the same buffer, and to slides/text.h
which shares struct text_cov with those tests.

+++ slides/text.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 struct text_cov {
-    uint16_t *data;
-    int       w, h;
+    uint8_t *data;
+    int      w, h;
 };
 
 struct text_cov text_rasterize(int W, int H, float font_size, _Bool sdf);
